Bug 1405739 P1 Don't expose internal redirect start/end/count through performance timing API. r=valentin
UltraBlame original commit: 6c5488ade824e460bfe19d4b347b78cfa8960a79

+++ dom/performance/PerformanceTiming.h
@@ -141,7 +141,7 @@ public:
     return GetDOMTiming()->GetUnloadEventEnd();
   }
 
-  uint16_t GetRedirectCount() const;
+  uint8_t GetRedirectCount() const;
 
   
   
@@ -285,7 +285,7 @@ private:
   TimeStamp mCacheReadStart;
   TimeStamp mResponseEnd;
   TimeStamp mCacheReadEnd;
-  uint16_t mRedirectCount;
+  uint8_t mRedirectCount;
   bool mTimingAllowed;
   bool mAllRedirectsSameOrigin;
   bool mInitialized;
